# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vertex Finance OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /subcategories:
    get:
      operationId: listSubcategories
      tags: [categories]
      summary: List subcategories
      responses:
        "200":
          description: Subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subcategory"
    post:
      operationId: createSubcategory
      tags: [categories]
      summary: Create subcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubcategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"

  # ─── TRANSACTIONS ────────────────────────────────────────────────────────────
...
  /monthly-plans:
    get:
      operationId: listMonthlyPlans
      tags: [monthly-plans]
      summary: List monthly plans
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyPlan"
    post:
      operationId: createMonthlyPlan
      tags: [monthly-plans]
      summary: Create monthly plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonthlyPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyPlan"
  /monthly-plans/{id}:
    put:
      operationId: updateMonthlyPlan
      tags: [monthly-plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonthlyPlanInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyPlan"

  # ─── ASSETS (PATRIMÔNIO) ─────────────────────────────────────────────────────
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/monthly-chart:
    get:
      operationId: getMonthlyChart
      tags: [dashboard]
      summary: Get monthly income vs expense chart data
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyChartItem"
  /dashboard/category-chart:
    get:
      operationId: getCategoryChart
      tags: [dashboard]
      summary: Get expense breakdown by category
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Category chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryChartItem"

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
  /reports/expenses-by-category:
    get:
      operationId: getReportExpensesByCategory
      tags: [reports]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expenses by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryReportItem"
  /reports/monthly-evolution:
    get:
      operationId: getReportMonthlyEvolution
      tags: [reports]
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly evolution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyEvolutionItem"
  /reports/top-expenses:
    get:
      operationId: getReportTopExpenses
      tags: [reports]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /reports/planned-vs-realized:
    get:
      operationId: getReportPlannedVsRealized
      tags: [reports]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Planned vs realized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannedVsRealizedItem"
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        group:
          type: string
          enum: [fixed, variable, leisure, investment_lp, investment_cp, other]
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, name, type, isActive]
    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        group:
          type: string
          enum: [fixed, variable, leisure, investment_lp, investment_cp, other]
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [name, type]
    Subcategory:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
      required: [id, categoryId, name, isActive]
    CreateSubcategoryInput:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
      required: [categoryId, name]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        competenceDate:
          type: string
          format: date
        movementDate:
          type: string
          format: date
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        subcategoryId:
          type: integer
          nullable: true
        subcategoryName:
          type: string
          nullable: true
        type:
          type: string
          enum: [income, expense, transfer]
        paymentMethod:
          type: string
          nullable: true
        description:
          type: string
        amount:
          type: number
        accountId:
          type: integer
          nullable: true
        accountName:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, paid, received]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, competenceDate, movementDate, type, description, amount, status, createdAt]
...
    MonthlyPlan:
      type: object
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        plannedIncome:
          type: number
        plannedExpense:
          type: number
        actualIncome:
          type: number
        actualExpense:
          type: number
        notes:
          type: string
          nullable: true
      required: [id, month, year, categoryId, categoryName, plannedIncome, plannedExpense, actualIncome, actualExpense]
    CreateMonthlyPlanInput:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        categoryId:
          type: integer
        plannedIncome:
          type: number
        plannedExpense:
          type: number
        notes:
          type: string
          nullable: true
      required: [month, year, categoryId, plannedIncome, plannedExpense]
...
    DashboardSummary:
      type: object
      properties:
        monthBalance:
          type: number
        totalIncome:
          type: number
        totalExpenses:
          type: number
        monthResult:
          type: number
        savingsPercentage:
          type: number
        totalAssets:
          type: number
        totalDebts:
          type: number
        totalReceivables:
          type: number
        netWorth:
          type: number
      required: [monthBalance, totalIncome, totalExpenses, monthResult, savingsPercentage, totalAssets, totalDebts, totalReceivables, netWorth]
    MonthlyChartItem:
      type: object
      properties:
        month:
          type: integer
        monthName:
          type: string
        income:
          type: number
        expenses:
          type: number
        result:
          type: number
      required: [month, monthName, income, expenses, result]
    CategoryChartItem:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        amount:
          type: number
        percentage:
          type: number
        color:
          type: string
          nullable: true
      required: [categoryId, categoryName, amount, percentage]
    CategoryReportItem:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        group:
          type: string
          nullable: true
        totalAmount:
          type: number
        transactionCount:
          type: integer
        percentage:
          type: number
      required: [categoryId, categoryName, totalAmount, transactionCount, percentage]
    MonthlyEvolutionItem:
      type: object
      properties:
        month:
          type: integer
        monthName:
          type: string
        income:
          type: number
        expenses:
          type: number
        result:
          type: number
        cumulativeResult:
          type: number
      required: [month, monthName, income, expenses, result, cumulativeResult]
    PlannedVsRealizedItem:
      type: object
      properties:
        groupId:
          type: integer
        groupName:
          type: string
        plannedAmount:
          type: number
        realizedAmount:
          type: number
        difference:
          type: number
        percentageUsed:
          type: number
        status:
          type: string
          enum: [on_track, over, under]
      required: [groupId, groupName, plannedAmount, realizedAmount, difference, percentageUsed, status]