# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vertex Finance OS API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  monthBalance: number;
  totalIncome: number;
  totalExpenses: number;
  monthResult: number;
  savingsPercentage: number;
  totalAssets: number;
  totalDebts: number;
  totalReceivables: number;
  netWorth: number;
}