# lib/api-zod/src/generated/types/monthlyEvolutionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vertex Finance OS API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyEvolutionItem {
  month: number;
  monthName: string;
  income: number;
  expenses: number;
  result: number;
  cumulativeResult: number;
}